// Amlak/CMS/CMS/src/app/shared/models/enums.ts
export enum DocumentType {
  PromiseLetter = 1,
  HasDocument = 2,
  Advocacy = 3,
}

export enum CustomerType {
  Tenant = 1,
  Owner = 2,
}

export enum ContractType {
  IntentionToSell = 1,
  IntentionToRent = 2,
  IntentionOfFullMortgage = 3,
  IntentionToMortgageAndRent = 4,
  IntentionToBuy = 5,
}

export enum AssetType {
  Image = 1,
  Video = 2,
  Other = 3,
}

export enum EstateType {
  Villa = 1,
  Official = 2,
  Area = 3,
  Commercial = 4,
  Apartment = 5,
  Garden = 6,
}